indian-market-advisor/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.1
python-dotenv==1.0.1
yfinance==0.2.50
feedparser==6.0.11
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
anthropic==0.40.0
apscheduler==3.10.4
httpx==0.28.0
